requirements.txt
allure-pytest==2.7.1
allure-python-commons==2.7.1
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
idna==2.8
image==1.5.27
importlib-metadata==0.19
more-itertools==7.2.0
packaging==19.1
Pillow==6.1.0
pluggy==0.12.0
py==1.8.0
pyparsing==2.4.2
pytest==5.0.1
pytz==2019.2
requests==2.22.0
selenium==3.141.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
wcwidth==0.1.7
zipp==0.5.2
